## 4.2.1.0 (#66)
* Add lasso tool to support region based imagery download
* Add support for KML icon scaling
* Sort icons by distance from camera to mitigate artifacts when map is tilted
* Reset expiration on certificate when credentials are re-entered
* Pull expiration from certificate if it's specified
* Reimplement KML parsing for network link import
* Fix GDAL band request parameters for monochrome imagery with alpha
* Add additional vehicle models
* Show nested child KML network links in Overlay Manager
* Allow user to modify child KML network link download intervals
* Handle KML altitude mode specified as `gx:altitudeMode` element
* Add support for open polyline extrusions
* Add extrusion support for negative heights (extrude down from base)
* Various bug fixes

+++ atak/ATAK/app/src/main/java/com/atakmap/android/vehicle/model/VehicleModelCache.java
@@ -236,7 +236,8 @@ public class VehicleModelCache {
                 return;
 
             // Delete icon cache if we're updating
-            FileSystemUtils.delete(ICON_DIR);
+            if (ICON_DIR.exists())
+                FileSystemUtils.delete(ICON_DIR);
         }
 
         // Read categories we need to load from
@@ -306,7 +307,7 @@ public class VehicleModelCache {
             String name = vehicle.getString("name");
             String path = vehicle.getString("file");
             File vehFile = new File(dir, path);
-            if (updating) {
+            if (updating && vehFile.exists()) {
                 // File is no longer up to date - delete so we can load from
                 // assets later
                 FileSystemUtils.delete(vehFile);
